+++ src/test/java/woowacrew/rss/utils/RssReaderTest.java
@@ -1,8 +1,14 @@
-package woowacrew.rss.util;
+package woowacrew.rss.utils;
 
 import org.junit.jupiter.api.Test;
+import org.springframework.core.io.ClassPathResource;
+import woowacrew.rss.domain.FeedArticle;
 import woowacrew.rss.exception.InvalidXmlException;
 
+import java.io.IOException;
+import java.util.List;
+
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
@@ -18,5 +24,14 @@ class RssReaderTest {
     void xml형식이_맞다면_에러없이_통과하는지_테스트() {
         String url = "https://vsh123.github.io/feed.xml";
         assertDoesNotThrow(() -> new RssReader(url));
+    }
+
+    @Test
+    void feedArticles를_잘_생성하는지_테스트() throws IOException {
+        ClassPathResource classPathResource = new ClassPathResource("feed.xml");
+        RssReader rssReader = new RssReader(classPathResource.getURL().toString());
+
+        List<FeedArticle> feedArticles = rssReader.getFeedArticle();
+        assertThat(feedArticles.size()).isEqualTo(3);
     }
 }
